# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profiles": {
      "get": {
        "operationId": "ProfileController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List profiles with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilesListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "operationId": "ProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}": {
      "get": {
        "operationId": "ProfileController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "operationId": "ProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "delete": {
        "operationId": "ProfileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}/document-sources": {
      "get": {
        "operationId": "ProfileController_listSources",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List document sources assigned to profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSourcesListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "operationId": "ProfileController_assignSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDocumentSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned document source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}/document-sources/{documentId}": {
      "delete": {
        "operationId": "ProfileController_unassignSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unassign result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}/task": {
      "post": {
        "operationId": "ProfileController_assignTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned task result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTaskDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "get": {
        "operationId": "ProfileController_getTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get assigned task for profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTaskDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "delete": {
        "operationId": "ProfileController_unassignTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unassign task from profile",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}/run": {
      "post": {
        "operationId": "ProfileController_runAggregate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enqueue aggregate analysis job for profile",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "jobId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Trade API",
    "description": "Auto-generated OpenAPI documentation for Trade backend",
    "version": "1.0.0",
    "contact": {
      "name": "API",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Profile numeric identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Profile name",
            "example": "Default profile"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "description": "Optional profile description",
            "example": "My primary research profile"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp (ISO 8601)",
            "example": "2025-01-01T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp (ISO 8601)",
            "example": "2025-01-02T08:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ]
      },
      "ProfilesListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Returned profiles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of profiles matching the query",
            "example": 123
          },
          "limit": {
            "type": "number",
            "description": "Page size (limit) applied to the query",
            "example": 50
          },
          "offset": {
            "type": "number",
            "description": "Offset applied to the query",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "ProfileCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Profile name"
          },
          "description": {
            "type": "object",
            "description": "Profile description"
          }
        },
        "required": [
          "name"
        ]
      },
      "ProfileUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Profile name"
          },
          "description": {
            "type": "object",
            "description": "Profile description"
          }
        }
      },
      "DocumentSourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Link identifier within the profile",
            "example": 10
          },
          "documentId": {
            "type": "string",
            "format": "uuid",
            "description": "Source document UUID"
          }
        },
        "required": [
          "id",
          "documentId"
        ]
      },
      "DocumentSourcesListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Assigned document sources",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentSourceDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of assigned sources",
            "example": 12
          },
          "limit": {
            "type": "number",
            "description": "Page size (limit) applied to the query",
            "example": 50
          },
          "offset": {
            "type": "number",
            "description": "Offset applied to the query",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "AssignDocumentSourceDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "description": "UUID of document to assign",
            "format": "uuid"
          }
        },
        "required": [
          "documentId"
        ]
      },
      "AssignTaskDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "number",
            "description": "Numeric id of task to assign",
            "example": 1
          }
        },
        "required": [
          "taskId"
        ]
      },
      "ProfileTaskDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "object",
            "description": "Currently assigned task id",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "taskId"
        ]
      },
...
    }
  }
}